+++ CHANGELOG.rst
@@ -3,6 +3,12 @@
 Changelog
 ---------
 
+Unreleased (2018-10-01)
++++++++++++++++++++++++
+
+* Remove rackspace package dependency `rfc6266_parser`.
+* Each driver's package dependencies are now optional (`#4 <https://github.com/scottwernervt/cloudstorage/issues/4>`_).
+
 0.6 (2018-07-24)
 ++++++++++++++++